finish 3d waypoint rendering

+++ common/src/main/java/xyz/wagyourtail/minimap/waypoint/Waypoint.java
@@ -1,6 +1,6 @@
 package xyz.wagyourtail.minimap.waypoint;
 
-public record Waypoint(int posX, int posY, int posZ, int colR, int colG, int colB, String name, String[] groups,
+public record Waypoint(int posX, int posY, int posZ, byte colR, byte colG, byte colB, String name, String[] groups,
                        String[] levels) {
 
 //    public static Waypoint deserialize(String inp) {
